requirements.txt
pytrends
duckduckgo-search
google-generativeai
twikit
jinja2
